requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiokafka==0.11.0
aiomysql==0.2.0
aiosignal==1.4.0
alibabacloud-devops20210625==4.23.2
alibabacloud-tea==0.4.3
alibabacloud_credentials==0.3.6
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.11
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.3
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncmy==0.2.9
atlastk==0.13.5
attrs==24.3.0
cachetools==5.3.1
certifi==2024.8.30
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
colorlog==6.7.0
cos_python_sdk_v5==1.9.38
crcmod==1.7
cryptography==43.0.0
Cython==3.1.4
ecdsa==0.19.1
et_xmlfile==2.0.0
Faker==19.3.1
fastapi==0.115.0
fastapi-pagination==0.14.1
filetype==1.2.0
frozenlist==1.7.0
greenlet==3.2.4
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.3
jmespath==0.9.5
kafka-python==2.0.2
lz4==4.4.4
MarkupSafe==3.0.3
multidict==6.6.4
nacos-sdk-python==0.1.14
numpy==1.26.4
openpyxl==3.1.5
oss2==2.18.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pdoc==14.6.1
pluggy==1.6.0
ply==3.11
prometheus_client==0.20.0
propcache==0.2.1
protobuf==4.25.3
psutil==7.1.0
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.20.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
PyJWT==2.7.0
pymongo==3.12.3
PyMySQL==1.0.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rsa==4.9.1
schedule==1.2.1
sentry-sdk==0.14.4
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.37.2
thriftpy2==0.5.2
toml==0.10.2
tortoise==0.1.1
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
urllib3==1.26.15
uvicorn==0.30.6
win32_setctime==1.2.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==1.0.2
yarl==1.18.3
schema==0.7.8

asyncio-mqtt==0.16.2
distro==1.9.0
docstring_parser==0.17.0
grpcio-status==1.62.1
h2==4.3.0
hpack==4.1.0
httplib2==0.31.0
hyperframe==6.1.0
jiter==0.11.1
jsonpath-ng==1.7.0
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
midscene-python==0.1.1
networkx==3.5
openai==1.109.1
opencv-python==4.11.0.86
outcome==1.3.0.post0
paho-mqtt==2.1.0
playwright==1.55.0
proto-plus==1.26.1
pure-python-adb==0.3.0.dev0
pyasn1_modules==0.4.2
pyee==13.0.0
pyparsing==3.2.5
PySocks==1.7.1
rich==14.2.0

shellingham==1.5.4

sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
uritemplate==4.2.0
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.2.0

pillow==10.4.0